+++ tests/Foundation/Http/KernelTest.php
@@ -29,6 +29,7 @@ class KernelTest extends TestCase
         $kernel = new Kernel($this->getApplication(), $this->getRouter());
 
         $this->assertEquals([
+            \Illuminate\Foundation\Http\Middleware\HandlePrecognitiveRequests::class,
             \Illuminate\Cookie\Middleware\EncryptCookies::class,
             \Illuminate\Session\Middleware\StartSession::class,
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,